refactor: simple provider state
