+++ packages/plugins/Flow/src/state/Provider.ts
@@ -1,8 +1,6 @@
-import { getEnumAsArray } from '@dimensiondev/kit'
 import type { Plugin } from '@masknet/plugin-infra'
-import { ProviderState, ProviderStorage } from '@masknet/plugin-infra/web3'
-import { EnhanceableSite, ExtensionSite } from '@masknet/shared-base'
-import { Account, isSameAddress } from '@masknet/web3-shared-base'
+import { ProviderState } from '@masknet/plugin-infra/web3'
+import { isSameAddress } from '@masknet/web3-shared-base'
 import {
     chainResolver,
     ChainId,
@@ -17,29 +15,12 @@ import { Providers } from './Connection/provider'
 
 export class Provider extends ProviderState<ChainId, ProviderType, NetworkType, Web3Provider, Web3> {
     constructor(override context: Plugin.Shared.SharedContext) {
-        const defaultValue: ProviderStorage<Account<ChainId>, ProviderType> = {
-            accounts: Object.fromEntries(
-                getEnumAsArray(ProviderType).map((x) => [
-                    x.value,
-                    {
-                        account: '',
-                        chainId: ChainId.Mainnet,
-                    },
-                ]),
-            ) as Record<ProviderType, Account<ChainId>>,
-            providers: Object.fromEntries(
-                [...getEnumAsArray(EnhanceableSite), ...getEnumAsArray(ExtensionSite)].map((x) => [
-                    x.value,
-                    ProviderType.Blocto,
-                ]),
-            ) as Record<EnhanceableSite | ExtensionSite, ProviderType>,
-        }
-
-        super(context, Providers, defaultValue, {
+        super(context, Providers, {
             isSameAddress,
             isValidChainId,
             isValidAddress,
             getDefaultChainId: () => ChainId.Mainnet,
+            getDefaultProviderType: () => ProviderType.Blocto,
             getDefaultNetworkType: () => NetworkType.Flow,
             getNetworkTypeFromChainId: (chainId: ChainId) =>
                 chainResolver.chainNetworkType(chainId) ?? NetworkType.Flow,
